Combine unit tests into existing integration test file
